Support logging to specified file
Currently, bats puts logs to a temporary file and removes the file
at the end of the test. This is not convenient for observing the logs.
This commit makes it possible for user to specify a log file and
observe the logs during test or for post test analyzing. A new option
named -l (or --logfile) is added for this function.