Add functionality to check the bash version
Bats has know issues with old versions of bash < 4.1 as explained
in issues #49 and #140.
These issues create confusion because bats runs the tests without
errors, but the results of the tests are wrong.
This patchs adds the capability to check for the bash version and
exit if it's not supported.